pyproject.toml
[project]
name = "deep-research-agent-strands"
version = "0.1.0"
description = "Deep Research Agent using AWS Strands Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "anthropic>=0.74.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "openai>=2.8.1",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "scikit-learn>=1.7.2",
    "pandas>=2.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/tools", "src/utils"]
